#1 use interface embedding for better mock generation
This fixes the previously failing duck typing.
The reason it previously failed is that Go's duck typing goes only one interface deep.
If my interface is a copy of another interface but has a return type which is also a copy of the interface of the original return type, duck typing does not work anymore.
As we only need our interface for mocking, we simply embedded the original interface into our interfaces.
This eases maintenance and is therefore a rather pragmatic solution.